requirements.txt
about-time==4.2.1
absl-py==2.1.0
aiohttp==3.9.5
aiohttp-socks==0.8.4
aiosignal==1.3.1
alive-progress==3.1.5
altgraph==0.17.4
annotated-types==0.6.0
anyio==4.2.0
appdirs==1.4.4
Appium-Python-Client==4.0.0
arrow==1.3.0
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.1.0
autoslot==2022.12.1
bcrypt==4.1.2
beautifulsoup4==4.12.2
beautifultable==1.1.0
binaryornot==0.4.4
blinker==1.7.0
blis==0.7.11
Brotli==1.1.0
bs4==0.0.2
cachetools==5.3.3
camelot-py==0.11.0
cartesia==1.0.6
catalogue==2.0.10
certifi==2023.11.17
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
clearview==0.2.1
click==8.1.7
cloudpathlib==0.16.0
colorama==0.4.6
comm==0.2.0
comtypes==1.3.1
confection==0.1.4
contourpy==1.2.0
cookiecutter==2.6.0
cryptography==42.0.2
customtkinter==5.2.2
cycler==0.12.1
cymem==2.0.8
darkdetect==0.8.0
debugpy==1.8.0
decorator==4.4.2
distro==1.9.0
elastic-transport==8.13.1
elasticsearch==8.14.0
es-core-news-sm @ https://github.com/explosion/spacy-models/releases/download/es_core_news_sm-3.7.0/es_core_news_sm-3.7.0-py3-none-any.whl#sha256=61e6e5530941f5880166855f09f60d7e6ba79ec1e8e45f96244bdb1eb169eb1d
et-xmlfile==1.1.0
executing==2.0.1
ExifRead==3.0.0
exrex==0.11.0
fake-useragent==1.5.1
fastapi==0.110.2
ffmpeg==1.4
ffmpeg-python==0.2.0
filelock==3.15.4
filetype==1.2.0
Flask==3.0.0
flatbuffers==24.3.25
flet==0.22.0
flet-core==0.22.0
flet-runtime==0.22.0
fonttools==4.47.2
frozenlist==1.4.1
future==1.0.0
geographiclib==2.0
geopy==2.4.1
get-nic==0.0.3
gevent==24.2.1
gitdb==4.0.11
gitignore_parser==0.1.11
GitPython==3.1.43
google-ai-generativelanguage==0.6.6
google-api-core==2.19.1
google-api-python-client==2.137.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
google-generativeai==0.7.2
googleapis-common-protos==1.63.2
googletransx==2.4.2
grapheme==0.6.0
greenlet==3.0.3
grequests==0.7.0
groq==0.9.0
grpcio==1.65.0
grpcio-status==1.62.2
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.17.3
httplib2==0.22.0
httptools==0.6.1
httpx==0.24.1
humanize==4.10.0
hyperframe==6.0.1
idna==3.6
ImageHash==4.3.1
imageio==2.34.0
imageio-ffmpeg==0.5.1
imbalanced-learn==0.12.3
inflection==0.5.1
iniconfig==2.0.0
ipykernel==6.27.1
ipython==8.18.1
iterators==0.2.0
itsdangerous==2.1.2
jax==0.4.30
jaxlib==0.4.30
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
jsonpickle==3.2.2
jupyter_client==8.6.0
jupyter_core==5.5.0
keyboard==0.13.5
kiwisolver==1.4.5
langcodes==3.3.0
langdetect==1.0.9
linkify-it-py==2.0.3
lxml==5.2.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.2
matplotlib-inline==0.1.6
mdit-py-plugins==0.4.1
mdurl==0.1.2
mediapipe==0.10.14
ml-dtypes==0.4.0
MouseInfo==0.1.3
moviepy==1.0.3
multidict==6.0.5
murmurhash==1.0.10
mysql==0.0.3
mysqlclient==2.2.4
nest-asyncio==1.5.8
netifaces==0.11.0
nltk==3.8.1
numpy==1.26.4
oauthlib==3.2.2
openai==1.35.9
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
openpyxl==3.1.2
opt-einsum==3.3.0
oracledb==2.2.1
outcome==1.3.0.post0
packaging==23.2
pandas==2.2.2
parso==0.8.3
pbkdf2==1.3
pdfminer.six==20240706
pefile==2023.2.7
pillow==10.2.0
platformdirs==4.1.0
pluggy==1.5.0
polling2==0.5.0
preshed==3.0.9
proglog==0.1.10
prompt-toolkit==3.0.41
proto-plus==1.24.0
protobuf==4.25.3
psutil==5.9.6
pure-eval==0.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
PyAudio==0.2.14
PyAutoGUI==0.9.54
pycparser==2.21
pydantic==2.7.1
pydantic_core==2.18.2
pyfiglet==1.0.2
PyGetWindow==0.0.9
Pygments==2.17.2
pyinstaller==6.4.0
pyinstaller-hooks-contrib==2024.1
PyMsgBox==1.0.9
PyMuPDF==1.24.7
PyMuPDFb==1.24.6
pyngrok==7.1.6
pyparsing==3.1.1
pypdf==4.2.0
PyPDF2==3.0.1
pyperclip==1.8.2
pypiwin32==223
pypng==0.20220715.0
pypresence==4.3.0
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
PyRect==0.2.0
PyScreeze==0.1.30
PySocks==1.7.1
pystray==0.19.5
pytest==8.2.2
pytest-asyncio==0.23.7
python-dateutil==2.8.2
python-dotenv==1.0.1
python-nmap==0.7.1
python-slugify==8.0.4
python-socks==2.5.0
python-whois==0.9.4
python-xlib==0.33
pyttsx3==2.90
pytube==15.0.0
pytweening==1.0.7
pytz==2024.1
PyWavelets==1.6.0
pywifi==1.1.12
pywin32==306
pywin32-ctypes==0.2.2
pyxhook==1.0.0
PyYAML==6.0.1
pyzmq==25.1.2
qrcode==7.4.2
regex==2023.12.25
repath==0.9.0
requests==2.31.0
requests-file==2.1.0
requests-futures==1.0.1
requests-toolbelt==1.0.0
rich==13.7.0
rich-argparse==1.5.2
rsa==4.9
scapy==2.5.0
scikit-learn==1.5.0
scipy==1.13.0
seaborn==0.13.2
selenium==4.16.0
setuptools==70.1.0
sherlock-project==0.14.4
six==1.16.0
smart-open==6.4.0
smmap==5.0.1
snatch @ file:///D:/DEV/local/FJRG2007/snatch
sniffio==1.3.0
sortedcontainers==2.4.0
sounddevice==0.4.7
soupsieve==2.5
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SpeechRecognition==3.10.1
SQLAlchemy==2.0.28
srsly==2.4.8
stack-data==0.6.3
starlette==0.37.2
stem==1.8.2
tabula-py==2.9.3
tabulate==0.9.0
termcolor==2.4.0
terminaltexteffects==0.11.0
text-unidecode==1.3
textual==0.71.0
thinc==8.2.3
threadpoolctl==3.5.0
tldextract==5.1.2
tornado==6.4
torrequest==0.1.0
tqdm==4.66.1
traitlets==5.14.0
trio==0.23.1
trio-websocket==0.11.1
typer==0.9.4
types-python-dateutil==2.8.19.20240106
typing_extensions==4.9.0
tzdata==2024.1
uc-micro-py==1.0.3
uritemplate==4.1.1
urllib3==2.1.0
uvicorn==0.29.0
wasabi==1.1.2
watchdog==4.0.0
watchfiles==0.21.0
wayback==0.4.5
wcwidth==0.2.12
weasel==0.3.4
websocket-client==1.7.0
websockets==11.0.3
Werkzeug==3.0.1
wheel==0.43.0
wifi==0.3.8
windows-curses==2.3.3
winspeech==1.0.1
wsproto==1.2.0
yarl==1.9.4
zope.event==5.0
zope.interface==6.4.post2
